configreader -> configparser following cpython/configparser
